Re #110 Attempting to add functionality that allows mobile app to select and activate a ghost entry when the entryId is specified in the URL. I.e.:
/mobile/index?entryId=24824

should look for a ghost entry with ID 24824 in the list coming back from
the server, and if it finds it, it should activate the ghost entry and
select the return result for immediate editing. This is done through a
new Javascript function activateGhostEntry(). See also displayEntry().
But, while this code works when it is called from the click handler, it
doesn't seem to fully work when called automatically when the list is
first displayed.

What happens is the ghost entry is activated on the server, but the UI
doesn't update and the entry isn't being edited after the call returns.